+++ src/wallet/fees.cpp
@@ -19,16 +19,7 @@ Amount GetRequiredFee(const CWallet &wallet, unsigned int nTxBytes) {
 
 Amount GetMinimumFee(const CWallet &wallet, unsigned int nTxBytes,
                      const CCoinControl &coin_control) {
-    Amount nFeeNeeded =
-        GetMinimumFeeRate(wallet, coin_control).GetFeeCeiling(nTxBytes);
-
-    // But always obey the maximum.
-    const Amount max_tx_fee = wallet.m_default_max_tx_fee;
-    if (nFeeNeeded > max_tx_fee) {
-        nFeeNeeded = max_tx_fee;
-    }
-
-    return nFeeNeeded;
+    return GetMinimumFeeRate(wallet, coin_control).GetFeeCeiling(nTxBytes);
 }
 
 CFeeRate GetRequiredFeeRate(const CWallet &wallet) {